internal: Consolidates changelog scripts under packages/changelog-generator
