+++ .gitignore
@@ -5,6 +5,7 @@
 
 # Production
 /build
+packages/changelog-generator/build
 
 # Generated files
 .docusaurus
