+++ package.json
@@ -8,7 +8,7 @@
     "build": "node scripts/build/build.mjs",
     "build:verbose": "VERBOSE=true node scripts/build/build.mjs",
     "build:nocache": "DISABLE_BUILD_CACHE=true node scripts/build/build.mjs",
-    "build:legacy": "pnpm run changelog:aggregate && pnpm run generate:rss && docusaurus build",
+    "build:legacy": "pnpm run changelog:compile:json && pnpm run changelog:compile:rss && docusaurus build",
     "cache:clear": "node scripts/build/cache.mjs clear",
     "cache:stats": "node scripts/build/cache.mjs stats",
     "cache:validate": "node scripts/build/validate.mjs",
@@ -18,14 +18,13 @@
     "serve": "docusaurus serve",
     "write-translations": "docusaurus write-translations",
     "write-heading-ids": "docusaurus write-heading-ids",
-    "changelog:new": "node scripts/create-changelog-entry.mjs",
-    "generate:rss": "node scripts/generate-rss-feed.mjs",
+    "changelog:entry:new": "pnpm -C packages/changelog-generator run entry:new",
+    "changelog:sync:all": "pnpm -C packages/changelog-generator run sync:all",
+    "changelog:compile:json": "pnpm -C packages/changelog-generator run compile:json",
+    "changelog:compile:rss": "pnpm -C packages/changelog-generator run compile:rss",
+    "changelog:preview": "pnpm -C packages/changelog-generator run preview",
+    "changelog:publish": "pnpm -C packages/changelog-generator run publish",
     "generate:redirects": "tsx scripts/generate-redirects.ts",
-    "changelog:aggregate": "node scripts/generate-changelog-data.mjs",
-    "changelog:build": "pnpm -C packages/changelog-generator run build",
-    "changelog:analyze": "pnpm -C packages/changelog-generator run analyze",
-    "changelog:apply": "pnpm -C packages/changelog-generator run apply",
-    "changelog:run": "pnpm -C packages/changelog-generator run run",
     "typecheck": "tsc",
     "format": "prettier --write .",
     "test": "vitest run",
@@ -88,11 +87,8 @@
     "docusaurus-plugin-openapi-docs": "^4.4.0",
     "docusaurus-plugin-search-glean": "^0.7.0",
     "docusaurus-theme-openapi-docs": "^4.4.0",
-    "feed": "^5.1.0",
-    "gray-matter": "^4.0.3",
     "js-yaml": "^4.1.0",
     "lucide-react": "^0.548.0",
-    "marked": "^16.4.1",
     "prism-react-renderer": "^2.3.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -117,16 +113,13 @@
     "@types/xml2js": "^0.4.14",
     "@vitejs/plugin-react": "^5.0.4",
     "@vitest/ui": "^3.2.4",
-    "chalk": "^5.6.2",
     "fast-levenshtein": "^3.0.0",
-    "inquirer": "^12.9.6",
     "jsdom": "^26.1.0",
     "npm-run-all": "^4.1.5",
     "open": "^10.1.2",
     "prettier": "^3.6.1",
     "sass": "^1.93.2",
     "sass-loader": "^16.0.5",
-    "slugify": "^1.6.6",
     "tsx": "^4.20.6",
     "typescript": "~5.9.2",
     "vitest": "^3.2.4",
