+++ packages/changelog-generator/README.md
@@ -17,8 +17,25 @@ This package analyzes releases and OpenAPI spec changes and opens PRs with rende
 ## Local usage
 
 ```bash
-pnpm -C packages/changelog-generator run run -- --dry-run
+pnpm changelog:sync:all -- --dry-run
 ```
 This writes preview files under `.changelog-preview-<timestamp>` and logs entries in debug mode when `DEBUG=changelog*`.
+
+## Available Commands
+
+### For Developers
+- `pnpm changelog:entry:new` - Manually create a changelog entry (interactive)
+
+### For CI/Automation
+- `pnpm changelog:sync:all` - Sync from all configured GitHub repos
+- `pnpm changelog:sync:all -- --dry-run` - Preview sync without creating PR
+
+### For Build System
+- `pnpm changelog:compile:json` - Compile .md entries → changelog.json
+- `pnpm changelog:compile:rss` - Generate RSS feed from JSON
+
+### Advanced (Two-Step CI)
+- `pnpm changelog:preview` - Analyze GitHub releases → output JSON
+- `pnpm changelog:publish` - Publish entries from JSON (creates PR)
