+++ packages/changelog-generator/package.json
@@ -7,10 +7,15 @@
     "changelog-generator": "build/index.js"
   },
   "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "run": "pnpm run build && node build/index.js run",
-    "analyze": "pnpm run build && node build/index.js analyze",
-    "apply": "pnpm run build && node build/index.js apply"
+    "build": "tsc -p tsconfig.json && chmod +x build/index.js",
+    "entry:new": "tsx src/index.ts entry-new",
+    "compile:json": "tsx src/index.ts compile-json",
+    "compile:rss": "tsx src/index.ts compile-rss",
+    "sync:all": "tsx src/index.ts sync-all",
+    "preview": "tsx src/index.ts preview",
+    "publish": "tsx src/index.ts publish",
+    "build:data": "pnpm run compile:json",
+    "build:rss": "pnpm run compile:rss"
   },
   "dependencies": {
     "dotenv": "^16.4.5",
@@ -18,13 +23,23 @@
     "find-up": "^8.0.0",
     "js-yaml": "^4.1.0",
     "octokit": "^4.1.4",
-    "@gleanwork/api-client": "^0.13.0"
+    "@gleanwork/api-client": "^0.13.0",
+    "commander": "^12.1.0",
+    "inquirer": "^12.9.6",
+    "slugify": "^1.6.6",
+    "chalk": "^5.6.2",
+    "marked": "^16.4.1",
+    "gray-matter": "^4.0.3",
+    "feed": "^5.1.0"
   },
   "devDependencies": {
     "@octokit/types": "^13.10.0",
     "@types/debug": "^4.1.12",
     "@types/js-yaml": "^4.0.9",
     "@types/node": "^22.18.13",
-    "typescript": "~5.9.2"
+    "@types/inquirer": "^9.0.7",
+    "@types/marked": "^6.0.0",
+    "typescript": "~5.9.2",
+    "tsx": "^4.20.6"
   }
 }
